mesa: Allow mesa to compile without the x11 feature
This also means that, to still provide a useful result, we now compile
ang package the EGL and GLES2 libraries. While at it, we compile a
fairly recent library, libgbm, needed for Weston, a wayland compositor.

Without X, the only EGL platform supported is drm.

There's a small bug in 8.0.4 that prevents "make install" to run, it
still tries to install libGL, so we grossly patch the build system to
not error out when trying to install the non existing libGL.

This patch also introduces a virtual/libgles2 package people can use
when depending on OpenGL ES 2.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>