mesa: Bump the mesa _git recipes to somewhere in 8.1.0
The reasoning here is that I want to build Weston, a wayland compositor,
that needs mesa master. However, wayland needs current mesa master and I
cannot just override the git version as between 8.0.4 and 8.1.0 mesa
changed its build system from autoconf/custonm makefiles to
autoconf/automake.

I could have overriden the recipes in my distro layer but decided that
it could live like that as it prepares the mesa recipes for the jump as
well.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>