initramfs: Make mkdir not fail
This is first step to "robustify" the initramfs init script. Further
down the line, I'd like to have init spawn a shell when something fails
(by installing a trap) so one can debug what's wrong more easily.

This patch make "mkdir foo" not fail if foo already exists.

Note that if we are sure that these directory are in the initramfs, we
could just remove those lines.